Refactor Problems 006-031: Separate execution code from algorithm code
