+++ problems/problem_007.py
@@ -10,7 +10,6 @@ Answer: 104743
 """
 
 import math
-import time
 
 
 def solve_naive(n: int) -> int:
@@ -137,110 +136,3 @@ def is_prime_optimized(num: int) -> bool:
         i += 6
 
     return True
-
-
-def test_solutions() -> None:
-    """テストケースで解答を検証"""
-    test_cases = [
-        (1, 2),  # 最初の素数は2
-        (2, 3),  # 2番目の素数は3
-        (3, 5),  # 3番目の素数は5
-        (4, 7),  # 4番目の素数は7
-        (5, 11),  # 5番目の素数は11
-        (6, 13),  # 6番目の素数は13（問題例）
-        (10, 29),  # 10番目の素数は29
-        (100, 541),  # 100番目の素数は541
-    ]
-
-    print("=== テストケース ===")
-    for n, expected in test_cases:
-        result_naive = solve_naive(n)
-        result_optimized = solve_optimized(n)
-        result_math = solve_mathematical(n)
-
-        print(f"n = {n}")
-        print(f"  Expected: {expected}")
-        print(f"  Naive: {result_naive} {'✓' if result_naive == expected else '✗'}")
-        print(
-            f"  Optimized: {result_optimized} {'✓' if result_optimized == expected else '✗'}"
-        )
-        print(
-            f"  Mathematical: {result_math} {'✓' if result_math == expected else '✗'}"
-        )
-        print()
-
-
-def main() -> None:
-    """メイン関数"""
-    n = 10001
-
-    print("=== Problem 007: 10001st prime ===")
-    print(f"Finding the {n}st prime number")
-    print()
-
-    # テストケース
-    test_solutions()
-
-    # 本問題の解答
-    print("=== 本問題の解答 ===")
-
-    # 各解法の実行時間測定
-    start_time = time.time()
-    result_naive = solve_naive(n)
-    naive_time = time.time() - start_time
-
-    start_time = time.time()
-    result_optimized = solve_optimized(n)
-    optimized_time = time.time() - start_time
-
-    start_time = time.time()
-    result_math = solve_mathematical(n)
-    math_time = time.time() - start_time
-
-    print(f"素直な解法: {result_naive:,} (実行時間: {naive_time:.6f}秒)")
-    print(f"最適化解法: {result_optimized:,} (実行時間: {optimized_time:.6f}秒)")
-    print(f"数学的解法: {result_math:,} (実行時間: {math_time:.6f}秒)")
-    print()
-
-    # 結果の検証
-    if result_naive == result_optimized == result_math:
-        print(f"✓ 解答: {result_optimized:,}")
-    else:
-        print("✗ 解答が一致しません")
-        print(f"  Naive: {result_naive}")
-        print(f"  Optimized: {result_optimized}")
-        print(f"  Mathematical: {result_math}")
-        return
-
-    # パフォーマンス比較
-    print("=== パフォーマンス比較 ===")
-    fastest_time = min(naive_time, optimized_time, math_time)
-    print(f"素直な解法: {naive_time / fastest_time:.2f}x")
-    print(f"最適化解法: {optimized_time / fastest_time:.2f}x")
-    print(f"数学的解法: {math_time / fastest_time:.2f}x")
-
-    # 詳細な計算過程の表示
-    print("\n=== 計算過程の詳細 ===")
-
-    # 最初の20個の素数を表示
-    print("最初の20個の素数:")
-    primes_list = []
-    for i in range(1, 21):
-        primes_list.append(str(solve_optimized(i)))
-    print(", ".join(primes_list))
-
-    # アルゴリズムの説明
-    print("\n=== アルゴリズムの説明 ===")
-    print("素直な解法: 各数を順次チェックして素数判定")
-    print("最適化解法: エラトステネスの篩で効率的に素数を生成")
-    print("数学的解法: 6k±1の形の数のみをチェックして効率化")
-
-    # 素数定理による近似
-    approx_nth_prime = n * math.log(n)
-    print(f"\n素数定理による{n}番目の素数の近似値: {approx_nth_prime:.0f}")
-    print(f"実際の値: {result_optimized}")
-    print(f"誤差: {abs(result_optimized - approx_nth_prime):.0f}")
-
-
-if __name__ == "__main__":
-    main()
