+++ problems/problem_008.py
@@ -10,7 +10,6 @@ What is the value of this product?
 Answer: 23514624000
 """
 
-import time
 from functools import reduce
 from operator import mul
 
@@ -155,117 +154,3 @@ def get_max_product_sequence(adjacent_digits: int = 13) -> tuple[str, int]:
             max_sequence = sequence
 
     return max_sequence, max_product
-
-
-def test_solutions() -> None:
-    """テストケースで解答を検証"""
-    test_cases = [
-        (1, 9),  # 1桁の最大値は9
-        (2, 81),  # 2桁の最大積（例：9×9=81）
-        (4, 5832),  # 問題例：4桁の場合は5832
-        (13, 23514624000),  # 本問題：13桁の場合
-    ]
-
-    print("=== テストケース ===")
-    for digits, expected in test_cases:
-        if digits <= len(THOUSAND_DIGIT_NUMBER):
-            result_naive = solve_naive(digits)
-            result_optimized = solve_optimized(digits)
-            result_math = solve_mathematical(digits)
-
-            print(f"Adjacent digits = {digits}")
-            print(f"  Expected: {expected}")
-            print(f"  Naive: {result_naive} {'✓' if result_naive == expected else '✗'}")
-            print(
-                f"  Optimized: {result_optimized} {'✓' if result_optimized == expected else '✗'}"
-            )
-            print(
-                f"  Mathematical: {result_math} {'✓' if result_math == expected else '✗'}"
-            )
-            print()
-
-
-def main() -> None:
-    """メイン関数"""
-    adjacent_digits = 13
-
-    print("=== Problem 008: Largest product in a series ===")
-    print(f"Finding the largest product of {adjacent_digits} adjacent digits")
-    print("In the 1000-digit number")
-    print()
-
-    # テストケース
-    test_solutions()
-
-    # 本問題の解答
-    print("=== 本問題の解答 ===")
-
-    # 各解法の実行時間測定
-    start_time = time.time()
-    result_naive = solve_naive(adjacent_digits)
-    naive_time = time.time() - start_time
-
-    start_time = time.time()
-    result_optimized = solve_optimized(adjacent_digits)
-    optimized_time = time.time() - start_time
-
-    start_time = time.time()
-    result_math = solve_mathematical(adjacent_digits)
-    math_time = time.time() - start_time
-
-    print(f"素直な解法: {result_naive:,} (実行時間: {naive_time:.6f}秒)")
-    print(f"最適化解法: {result_optimized:,} (実行時間: {optimized_time:.6f}秒)")
-    print(f"数学的解法: {result_math:,} (実行時間: {math_time:.6f}秒)")
-    print()
-
-    # 結果の検証
-    if result_naive == result_optimized == result_math:
-        print(f"✓ 解答: {result_optimized:,}")
-    else:
-        print("✗ 解答が一致しません")
-        print(f"  Naive: {result_naive}")
-        print(f"  Optimized: {result_optimized}")
-        print(f"  Mathematical: {result_math}")
-        return
-
-    # パフォーマンス比較
-    print("=== パフォーマンス比較 ===")
-    fastest_time = min(naive_time, optimized_time, math_time)
-    print(f"素直な解法: {naive_time / fastest_time:.2f}x")
-    print(f"最適化解法: {optimized_time / fastest_time:.2f}x")
-    print(f"数学的解法: {math_time / fastest_time:.2f}x")
-
-    # 詳細な計算過程の表示
-    print("\n=== 計算過程の詳細 ===")
-
-    # 最大積となるシーケンスを表示
-    max_sequence, max_product = get_max_product_sequence(adjacent_digits)
-    print(f"最大積となる{adjacent_digits}桁のシーケンス: {max_sequence}")
-    print(f"各桁: {' × '.join(max_sequence)}")
-    print(f"積: {max_product:,}")
-
-    # 問題例（4桁の場合）も表示
-    print("\n問題例（4桁の場合）:")
-    example_sequence, example_product = get_max_product_sequence(4)
-    print(f"最大積となる4桁のシーケンス: {example_sequence}")
-    print(f"各桁: {' × '.join(example_sequence)} = {example_product}")
-
-    # アルゴリズムの説明
-    print("\n=== アルゴリズムの説明 ===")
-    print("素直な解法: 全ての隣接するシーケンスをチェックして積を計算")
-    print("最適化解法: ゼロを含むシーケンスをスキップするスライディングウィンドウ")
-    print("数学的解法: reduce関数を使用した効率的な積計算とゼロスキップ")
-
-    # 1000桁の数値に関する統計
-    zero_count = THOUSAND_DIGIT_NUMBER.count("0")
-    total_digits = len(THOUSAND_DIGIT_NUMBER)
-    print("\n=== 1000桁数値の統計 ===")
-    print(f"総桁数: {total_digits}")
-    print(f"ゼロの個数: {zero_count}")
-    print(f"ゼロの割合: {zero_count / total_digits * 100:.1f}%")
-    print(f"最大桁: {max(THOUSAND_DIGIT_NUMBER)}")
-    print(f"最小桁: {min(THOUSAND_DIGIT_NUMBER)}")
-
-
-if __name__ == "__main__":
-    main()
