+++ problems/problem_025.py
@@ -98,28 +98,3 @@ def solve_optimized(target_digits: int = 1000) -> int:
         index += 1
 
     return index
-
-
-def main() -> None:
-    """Main function to run and compare solutions."""
-    import time
-
-    target_digits = 1000
-
-    print("Solving Problem 025...")
-
-    # --- Naive Solution ---
-    start_time = time.time()
-    naive_answer = solve_naive(target_digits)
-    naive_time = time.time() - start_time
-    print(f"Naive solution: {naive_answer} (took {naive_time:.6f} seconds)")
-
-    # --- Optimized Solution ---
-    start_time = time.time()
-    optimized_answer = solve_optimized(target_digits)
-    optimized_time = time.time() - start_time
-    print(f"Optimized solution: {optimized_answer} (took {optimized_time:.6f} seconds)")
-
-
-if __name__ == "__main__":
-    main()
