+++ problems/problem_029.py
@@ -43,28 +43,3 @@ def solve_optimized(limit: int = 100) -> int:
     Therefore, the optimized solution is the same as the naive one.
     """
     return solve_naive(limit)
-
-
-def main() -> None:
-    """Main function to run and compare solutions."""
-    import time
-
-    limit = 100
-
-    print(f"Solving Problem 029 for limit = {limit}...")
-
-    # --- Naive Solution ---
-    start_time = time.time()
-    naive_answer = solve_naive(limit)
-    naive_time = time.time() - start_time
-    print(f"Naive solution: {naive_answer} (took {naive_time:.6f} seconds)")
-
-    # --- Optimized Solution ---
-    start_time = time.time()
-    optimized_answer = solve_optimized(limit)
-    optimized_time = time.time() - start_time
-    print(f"Optimized solution: {optimized_answer} (took {optimized_time:.6f} seconds)")
-
-
-if __name__ == "__main__":
-    main()
