+++ problems/problem_030.py
@@ -40,23 +40,3 @@ def solve(power: int = 5) -> int:
             total_sum += i
 
     return total_sum
-
-
-def main() -> None:
-    """Main function to run and display the solution."""
-    import time
-
-    power = 5
-
-    print(f"Solving Problem 030 for power = {power}...")
-
-    start_time = time.time()
-    solution = solve(power)
-    elapsed_time = time.time() - start_time
-
-    print(f"Solution: {solution}")
-    print(f"Took {elapsed_time:.6f} seconds")
-
-
-if __name__ == "__main__":
-    main()
