+++ problems/problem_031.py
@@ -14,8 +14,6 @@ How many different ways can £2 be made using any number of coins?
 Answer: Check on Project Euler official site
 """
 
-import time
-
 
 def solve_naive(target: int = 200) -> int:
     """
@@ -54,52 +52,3 @@ def solve_optimized(target: int = 200) -> int:
             dp[amount] += dp[amount - coin]
 
     return dp[target]
-
-
-def test_solutions() -> None:
-    """テストケースで解答を検証"""
-    assert (
-        solve_optimized(5) == 4
-    )  # 5pを作る方法: 5p, 2p+2p+1p, 2p+1p+1p+1p, 1p+1p+1p+1p+1p
-    assert solve_optimized(10) == 11  # 10pを作る方法
-
-    # メイン問題のテスト（素直な解法は遅いので小さい値のみ）
-    assert solve_naive(10) == solve_optimized(10)
-
-    print("All tests passed!")
-
-
-def main() -> None:
-    """メイン関数"""
-    print("=== Problem 031: Coin sums ===")
-
-    # テスト実行
-    test_solutions()
-
-    # 実際の問題を解く（200p = £2）
-    start_time = time.time()
-    result_optimized = solve_optimized(200)
-    optimized_time = time.time() - start_time
-
-    print(f"Optimized solution: {result_optimized}")
-    print(f"Execution time (optimized): {optimized_time:.6f} seconds")
-
-    # 小さい値での比較
-    print("\nSmall test comparisons:")
-    for test_val in [5, 10, 20]:
-        start_time = time.time()
-        naive_result = solve_naive(test_val)
-        naive_time = time.time() - start_time
-
-        start_time = time.time()
-        opt_result = solve_optimized(test_val)
-        opt_time = time.time() - start_time
-
-        print(
-            f"Target {test_val}p: Naive={naive_result} ({naive_time:.6f}s), "
-            f"Optimized={opt_result} ({opt_time:.6f}s)"
-        )
-
-
-if __name__ == "__main__":
-    main()
